refactor: move maintenance tools and utility into the package
Move the maintenance tooling under the package (Phase 2, PR-06):
- `git mv holdings_maintenance src/pdsfile/holdings_maintenance`
- `git mv utility/show_opus_products.py src/pdsfile/tools/show_opus_products.py`
  (the now-empty `utility/` directory is removed)
- rename the two hyphenated modules to importable names (rename only):
  `re-validate.py` -> `re_validate.py`,
  `shelf-consistency-check.py` -> `shelf_consistency_check.py`

Minimal content edits required by the move (itemized):
- 11 `[project.scripts]` targets reprefixed `holdings_maintenance.…` ->
  `pdsfile.holdings_maintenance.…`.
- `pdsinfoshelf.py` and `re_validate.py`: drop the `REPO_ROOT` +
  `sys.path.insert` hack and its now-dead `from pathlib import Path`
  import; the intra-package imports become
  `from pdsfile.holdings_maintenance.pds3 import …`.
- `pds4infoshelf.py`: repoint `from holdings_maintenance.pds4 import
  pds4checksums` -> `from pdsfile.holdings_maintenance.pds4 import …`
  (a third cross-import site the plan's itemization did not name, but a
  necessary consequence of the move — it had no sys.path hack).
- `show_opus_products.py`: replace the two
  `from pdsfile.pds{3,4}file.tests.helper import PDS{3,4}_HOLDINGS_DIR`
  imports with direct `os.environ[...]` reads (the tests package leaves
  the distribution in PR-07); preload call sites unchanged.

Packaging: collapse discovery to `src`-only
(`packages.find where = ["src"]`, `include = ["pdsfile*"]`) now that the
tools live under `src/pdsfile/`.

Deviation from the literal spec (flagged for review): add three empty
`__init__.py` files (`holdings_maintenance/pds3`, `holdings_maintenance/pds4`,
`tools`). Without them setuptools' `find_packages` silently drops those
directories from a built wheel — the editable install masks it, but every
console-script entry point (all of which target
`pdsfile.holdings_maintenance.pds{3,4}.*`) plus the `show_opus_products`
tool would be missing from a real install. Verified by building a wheel
and confirming the modules are present.

Ratchet/target bookkeeping: reprefix the ruff per-file-ignores keys to
`src/pdsfile/...` (and the two renamed modules); drop `holdings_maintenance`
and `utility` from the ruff targets in `gen_ruff_ratchet.py` and
`run-all-checks.sh` (now covered by `src/pdsfile`).

Verification: `pip install -e .` succeeds; all 11 console scripts resolve
and `--help`; `python -m pdsfile.tools.show_opus_products --help` works;
ruff (ratchet), pyroma 10/10 and the API-freeze manifest pass; the full
`--mode ns` suite is 679 passed / 34 skipped / 0 failed against the
limited testing holdings — unchanged from the pre-move baseline.
