+++ src/pdsfile/holdings_maintenance/pds3/copy_all_except_metadata.sh
