+++ src/pdsfile/holdings_maintenance/pds3/pdsinfoshelf.py
@@ -12,7 +12,6 @@ import argparse
 import datetime
 import glob
 import os
-from pathlib import Path
 import pickle
 import re
 import shutil
@@ -22,10 +21,7 @@ from PIL import Image
 import pdslogger
 import pdsfile
 
-REPO_ROOT = Path(__file__).resolve().parent.parent.parent
-sys.path.insert(0, str(REPO_ROOT))
-
-from holdings_maintenance.pds3 import pdschecksums
+from pdsfile.holdings_maintenance.pds3 import pdschecksums
 
 # Holds log file directories temporarily, used by move_old_info()
 LOGDIRS = []
